Extract utility methods from `MembraneMCPServer` to new `MCPUtil` class for improved modularity and code reuse.
